+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [Unreleased]
+## [8.5.0-rc.1] - 2026-09-21
 
 ### Changed
 
@@ -1517,7 +1517,7 @@ _Stable release based on [0.1.0-rc.1]._
 
 - New changelog file.
 
-[Unreleased]: https://https://github.com/internetguru/laravel-user/compare/staging...dev
+[8.5.0-rc.1]: https://github.com/internetguru/laravel-user/releases/tag/v8.4.0
 [8.4.0]: https://https://github.com/internetguru/laravel-user/compare/v8.3.0...v8.4.0
 [8.4.0-rc.1]: https://github.com/internetguru/laravel-user/releases/tag/v8.3.0
 [8.3.0]: https://https://github.com/internetguru/laravel-user/compare/v8.2.2...v8.3.0